refactor: 파일 정리 및 폴더 이동
