+++ src/main/java/racingcar/common/constant/MessageConstants.java
@@ -1,4 +1,4 @@
-package racingcar.constant;
+package racingcar.common.constant;
 
 public class MessageConstants {
     public static final String INPUT_CAR_NAMES = "경주할 자동차 이름을 입력하세요. (이름은 쉼표(,) 기준으로 구분)";
@@ -11,6 +11,5 @@ public class MessageConstants {
     public static final String INVALID_CAR_NAME_ERROR = "잘못 된 자동차 이름입니다.";
     public static final String INVALID_COMMAND_ERROR = "잘못 된 커맨드를 입력입니다.";
     public static final String NEGATIVE_ITERATION_ERROR = "반복 횟수는 음수가 될 수 없습니다..";
-    public static final String FILTER_NOT_FOUND_ERROR = "필터를 찾을 수 없습니다.";
 
 }
