+++ src/main/java/racingcar/exception/InvalidCommandException.java
@@ -1,6 +1,6 @@
 package racingcar.exception;
 
-import racingcar.constant.MessageConstants;
+import racingcar.common.constant.MessageConstants;
 
 public class InvalidCommandException extends IllegalArgumentException {
     public InvalidCommandException() {
